kurento_bower_publish: check if tag exists prior to commit into the repository

+++ kurento_bower_publish.sh
@@ -33,6 +33,12 @@ COMMIT_MSG=$(echo $COMMIT_LOG|cut -d' ' -f2-)
 MESSAGE="Generated code from $KURENTO_PROJECT:$COMMIT_ID"
 VERSION=$(xmlstarlet sel -N x=http://maven.apache.org/POM/4.0.0 -t -v "/x:project/x:version" pom.xml)
 
+# Fail if a tag for that version already exists
+if [ ${CREATE_TAG} = true ] && [[ $VERSION != *-SNAPSHOT ]]; then
+  # If tag already exists terminate silently
+  git ls-remote --tags|grep -q "$VERSION" && exit 0
+fi
+
 # Checkout bower repo
 BOWER_DIR="bower_code"
 [ -d $BOWER_DIR ] && rm -rf $BOWER_DIR
@@ -75,9 +81,6 @@ fi
 # Check if a version has to be generated
 if [ ${CREATE_TAG} = true ] && [[ $VERSION != *-SNAPSHOT ]]; then
   echo "Add Tag version: $VERSION"
-
-  # If tag already exists terminate silently
-  git ls-remote --tags|grep -q "$VERSION" && exit 0
 
   # Add tag
   git tag $VERSION || exit 1
